会议数据配置

+++ foxconn-sw-tool-service/src/main/java/com/foxconn/sw/service/processor/GeneralProcessor.java
@@ -18,7 +18,7 @@ public class GeneralProcessor {
         generalVo.setScheduleTask(taskBusiness.getTaskCount());
         generalVo.setUnreadTask(taskBusiness.getUnReadTaskCount());
         generalVo.setCollaboration(taskBusiness.getCollaborationCount());
-        generalVo.setMeeting(meetingBusiness.getMeetingCount());
+        generalVo.setMeeting(meetingBusiness.getMeetingCount().intValue());
         return generalVo;
     }
